requirements.txt
fastapi
uvicorn
pan-aadhar-ocr
python-multipart
python-dotenv
requests